requirements.txt
# Web framework + server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# Database
aiosqlite<=0.17.0

# Platform collectors
tweepy>=4.14.0
asyncpraw>=7.8.0
google-api-python-client>=2.140.0
google-auth>=2.30.0

# Scheduling
apscheduler>=3.10.0

# Configuration
pyyaml>=6.0
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.0
